Responsive layout and Material UI elems

+++ src/site/components/AccountView/index.js
@@ -1,9 +1,14 @@
 import React, { Component, PropTypes } from "react";
+import s from "./AccountView.scss";
+import withStyles from "../../decorators/withStyles";
+import TextField from "material-ui/lib/text-field";
+import FlatButton from "material-ui/lib/flat-button";
 import BaseComponent from "../BaseComponent";
 import Header from "../Header";
 import Feedback from "../Feedback";
 import Footer from "../Footer";
 
+@withStyles(s)
 class AccountView extends BaseComponent {
   
   static contextTypes = {
@@ -13,6 +18,7 @@ class AccountView extends BaseComponent {
   constructor(props) {
     super(props);
     console.log("constr account", props);
+    this.showPwd = false;
     this.state = Object.assign(props.state, {
     });
   }
@@ -20,6 +26,11 @@ class AccountView extends BaseComponent {
   async fetchData(props = this.props) { 
 
     if (props.state.status === "deleted") {
+      return;
+    }
+
+    if (this.showPwd) {
+      this.showPwd = false;
       return;
     }
 
@@ -45,22 +56,56 @@ class AccountView extends BaseComponent {
     let formParams = {};
     formParams[event.target.name] = event.target.value;
     let account = Object.assign(this.state.account, formParams);
-    this.setState(Object.assign(this.state, { account, messages: { editStatus: "changed" } }));
-  
+    let state = { account: Object.assign({}, account) };
+    
+    if (event.target.name === "password") {
+      state = { account: Object.assign(account, { password: account.password }) };
+      this.setState(state);
+    } else {
+      delete state.account.password;
+      state.messages = { editStatus: "changed" };
+      this.setState(state);
+    }
+
   }
   
   _saveAccount() {
-    this.props.saveAccount(this.state.account)
+    const account = Object.assign({}, this.state.account);
+    delete account.password;
+    this.props.saveAccount(account)
   }
   
+  _savePassword() {
+    this.props.saveAccount({
+      uuid: this.state.account.uuid, 
+      password: this.state.account.password 
+    });
+  }
+
   _deleteAccount() {
-    this.props.deleteAccount(this.state.account.uuid)
+    if (window.confirm("Olet poistamassa tiliäsi. Tätä toimenpidettä ei ehkä ole mahdollista peruuttaa. Oletko varma, että haluat jatkaa?")) {
+      this.props.deleteAccount(this.state.account.uuid);
+    }
+  }
+
+  _togglePassword() {
+    this.showPwd = true;
+    const pwdContainer = document.getElementById("password");
+    const pwdField = pwdContainer.getElementsByTagName("input")[0];
+    const imageElem = pwdContainer.getElementsByTagName("i")[0];
+    if (pwdField.type === "password") {
+      pwdField.type = "text";
+      imageElem.innerHTML = "&#xE8F4;";
+    } else {
+      pwdField.type = "password";
+      imageElem.innerHTML = "&#xE8F5;";
+    }
   }
 
   render() {
      
     console.log("render account", this.props, this.state);
-    let { account, messages } = this.state;
+    let { account, messages, pwdSaved } = this.state;
     
     if (!account) {
       return null;
@@ -74,56 +119,120 @@ class AccountView extends BaseComponent {
     }
     
     let editStatus = null;
-    if (messages && messages.editStatus) {
+    if (messages && messages.editStatus && !pwdSaved) {
+      
+      let statusColor = {};
+      switch (messages.editStatus) {
+        case "changed":
+          statusColor = { color: "blue" };
+          break;
+        case "saved":
+          statusColor = { color: "green" };
+          break;
+        default:
+          statusColor = {};
+      }
+
       editStatus = (
-        <span>{messages.editStatus}</span>
+        <span style={statusColor}>{messages.editStatus}</span>
       );
     }
+
+    let pwdStatus = null;
+    if (pwdSaved) {
+      pwdStatus = (
+        <span style={{ color: "green" }}>Vaihdettu</span>
+      );
+    }
+    
+    let fullWidth = { width: "100%" };
+    let deleteButtonCss = { 
+      width: "100%",
+      backgroundColor: "#DC4E41" 
+    };
+
+    const showPwdCss = {
+      position: "absolute",
+      right: "0px",
+      top: "40px"
+    };
 
     return (
       <div>
         <Header auth={this.state.auth} />
         {formError}
-        <div>
-          <div>
-            Sähköposti
-            <input type="text" 
-              name="email" 
-              value={account.email}
-              onChange={this._handleInputChange.bind(this)} />
+        <div className={s.root}>
+          <div className={s.profile}>
+            <div className={s.profileTitle}>
+              Henkilötiedot 
+            </div>
+            <div className={s.profileRow}>
+              <div className={s.profileCell}>
+                <TextField style={fullWidth}
+                  name="email" 
+                  floatingLabelText="Sähköposti"
+                  value={account.email}
+                  onChange={this._handleInputChange.bind(this)} />
+              </div>
+              <div className={s.profileCell}>
+                <TextField style={fullWidth} 
+                  name="username" 
+                  floatingLabelText="Käyttäjätunnus"
+                  value={account.username}
+                  onChange={this._handleInputChange.bind(this)} />
+              </div>
+            </div>
+            <div className={s.profileRow}>
+              <div className={s.profileCell}>
+                <TextField style={fullWidth}
+                  name="firstName" 
+                  floatingLabelText="Etunimi"
+                  value={account.firstName}
+                  onChange={this._handleInputChange.bind(this)} />
+              </div>
+              <div className={s.profileCell}>
+                <TextField style={fullWidth} 
+                  name="lastName" 
+                  floatingLabelText="Sukunimi"
+                  value={account.lastName}
+                  onChange={this._handleInputChange.bind(this)} />
+              </div>
+            </div>
+            <div className={s.saveProfile}>
+              <div className={s.editStatus}>
+                {editStatus}
+              </div>
+              <div className={s.saveButton}>
+                <FlatButton onClick={() => this._saveAccount()} label="Tallenna" />
+              </div>
+            </div>
           </div>
-          <div>
-            Salasana
-            <input type="text" 
-              name="password" 
-              value={account.password}
-              onChange={this._handleInputChange.bind(this)} />
+          <div className={s.password}>
+            <div className={s.passwordTitle}>
+              Salasanan vaihto 
+            </div>
+            <div className={s.passwordInput} id="password">
+              <TextField style={fullWidth} 
+                type="password"
+                name="password" 
+                floatingLabelText="Uusi salasana"
+                value={account.password}
+                onChange={this._handleInputChange.bind(this)} />
+              <a href="#" onClick={() => this._togglePassword()} style={showPwdCss}>
+                <i className="material-icons">&#xE8F5;</i>
+              </a>
+            </div>
+            <div className={s.changePassword}>
+              <div className={s.pwdStatus}>
+                {pwdStatus}
+              </div>
+              <div className={s.changePwdButton}>
+                <FlatButton onClick={() => this._savePassword()} label="Vaihda" />
+              </div>
+            </div>
           </div>
-          <div>
-            Tunnus
-            <input type="text" 
-              name="username" 
-              value={account.username}
-              onChange={this._handleInputChange.bind(this)} />
-          </div>
-          <div>
-            Etunimi
-            <input type="text" 
-              name="firstName" 
-              value={account.firstName}
-              onChange={this._handleInputChange.bind(this)} />
-          </div>
-          <div>
-            Sukunimi
-            <input type="text" 
-              name="lastName" 
-              value={account.lastName}
-              onChange={this._handleInputChange.bind(this)} />
-          </div>
-          <div>
-            {editStatus}
-            <button onClick={() => this._saveAccount()}>Tallenna</button>
-            <button onClick={() => this._deleteAccount()}>Poista</button>
+          <div className={s.deleteAccount}>
+            <FlatButton style={deleteButtonCss} onClick={() => this._deleteAccount()} label="POISTA TILI" />
           </div>
         </div>
         <Feedback />
